Python: Separate AsyncLogClient from the LogClient
There really is no reason for those separate classes to live in the same
package.
